'solve' circular dependency, for now
