+++ hvcc/core/hv2ir/HeavyParser.py
@@ -30,7 +30,7 @@ from .HLangAdc import HLangAdc
 from .HLangBinop import HLangBinop
 from .HLangBiquad import HLangBiquad
 from .HLangDelay import HLangDelay
-from .HLangIf import HLangIf
+# from .HLangIf import HLangIf
 from .HeavyException import HeavyException
 from .HeavyIrObject import HeavyIrObject
 from .HeavyLangObject import HeavyLangObject
@@ -197,6 +197,32 @@ class HeavyParser:
             return LANG_CLASS_DICT[obj_type]
         else:
             return None
+
+
+class HLangIf(HeavyLangObject):
+    """ Translates HeavyLang object [if] to HeavyIR [if] or [if~].
+    """
+
+    def __init__(self, obj_type, args, graph, annotations=None):
+        HeavyLangObject.__init__(self, "if", args, graph,
+                                 num_inlets=2,
+                                 num_outlets=2,
+                                 annotations=annotations)
+
+    def reduce(self):
+        if self.has_inlet_connection_format(["cc", "_c", "c_", "__"]):
+            x = HeavyIrObject("__if", self.args)
+        elif self.has_inlet_connection_format("ff"):
+            # TODO(mhroth): implement this
+            x = HeavyParser.graph_from_file("./hvlib/if~f.hv.json")
+        elif self.has_inlet_connection_format("ii"):
+            # TODO(mhroth): implement this
+            x = HeavyParser.graph_from_file("./hvlib/if~i.hv.json")
+        else:
+            raise HeavyException("Unhandled connection configuration to object [if]: {0}".format(
+                self._get_connection_format(self.inlet_connections)))
+
+        return ({x}, self.get_connection_move_list(x))
 
 
 # NOTE(mhroth): these imports are at the end of the file in order to prevent
